+++ s/tools/explode_promise.ts
@@ -1,7 +1,7 @@
 
 export type PromiseParts = ReturnType<typeof explode_promise>
 
-/** @deprecated use `deferredPromise` instead */
+/** @deprecated use `deferPromise` instead */
 export function explode_promise<T>() {
 	let resolve!: (value: T) => void
 	let reject!: (reason: any) => void